// sourcecode/BTrak.Packages/snova-leave-management/leave-management-module/leave-management-components/src/lib/leavemanagement/models/leave-applicability-model.ts
export class LeaveApplicabilityModel {
    leaveApplicabilityId: string;
    leaveTypeId: string;
    roleId: any;
    roleIds: any;
    employeeTypeId: string;
    genderId: any;
    maritalStatusId: any;
    branchId: any;
    departmentId: any;
    employeeIds: any;
    employeeId: any;
    minExperienceInMonths: number;
    maxExperienceInMonths: number;
    maxLeaves: number;
    isArchived: boolean;
    timeStamp: any;
}